Move train entrypoint helpers to train_config.py (pure move)
Relocate the `build_trainer`, `get_validation_callback`,
`get_validate_stepper_callback`, `get_inference_callback`, the private
`_make_*_aggregator_factory` helpers, and the `AggregatorBuilder`/
`CoupledAggregatorBuilder` classes from the ace and coupled `train.py`
entrypoints into their respective `train_config.py`, next to the
`TrainConfig` they will be refactored onto. The function bodies are moved
verbatim — no logic changes — and re-exported from `train.py` so existing
importers and tests are unaffected. This isolates the mechanical move from
the behavioral refactor that follows.
